ajusto codigo para dividir logica cliente servidor
