+++ frontend/app-front/app/routes/movies.tsx
@@ -2,7 +2,7 @@ import { HeaderFull } from "components/ui/header-full";
 import { MovieCard } from "components/ui/movie-card";
 import { SidebarNav } from "components/ui/sidebar-nav";
 import { useLoaderData } from "react-router-dom";
-import { getMovies } from "services/api/movies";
+import { getMovies } from "services/api/flixy/server/movies";
 import type { Route } from "./+types/movies";
 
 interface Page<T> {
