+++ frontend/app-front/services/api/flixy/types/auth.ts
@@ -1,0 +1,11 @@
+export type LoginCredentials = {
+    username: FormDataEntryValue | null;
+    password: FormDataEntryValue | null;
+  };
+  
+  export type RegistrationData = {
+    name: FormDataEntryValue | null;
+    username: FormDataEntryValue | null;
+    email: FormDataEntryValue | null;
+    password: FormDataEntryValue | null;
+  };
