+++ frontend/app-front/services/api/flixy/types/overall.ts
@@ -1,0 +1,6 @@
+export interface ApiResponse {
+  error?: string | null
+  success?: string | null;
+  [key: string]: string | null | undefined;
+}
+
